% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{approximationSetUnknownGenotypeCombinations.control}
\alias{approximationSetUnknownGenotypeCombinations.control}
\title{Control function for the 'approximationSetUnknownGenotypeCombinations' function.}
\usage{
approximationSetUnknownGenotypeCombinations.control(simplifiedReturn = FALSE,
  levelsOfStutterRecursion = 2, numberOfThreads = 4, trace = FALSE,
  EAApproximationType = 2, numberOfSimulationsMH = 10000,
  suggestionMH = "random")
}
\arguments{
\item{simplifiedReturn}{Should the returned list be simplified (TRUE/FALSE)?}

\item{levelsOfStutterRecursion}{The number of layers used in the stutter recursion.}

\item{numberOfThreads}{The maximum number of threads allowed.}

\item{trace}{Show trace (TRUE/FALSE)?}

\item{EAApproximationType}{The type of EA approximation used to find the normalising constant.}

\item{numberOfSimulationsMH}{The number of simulations used for each marker of the Metropolis-Hastings algorithm.}

\item{suggestionMH}{The proposal distribution used in the Metropolis-Hastings algorithm.}
}
\value{
A list of default parameters.
}
\description{
A control function setting default parameters for the \link{approximationSetUnknownGenotypeCombinations} function.
}
