% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{approximationSetUnknownGenotypeCombinations}
\alias{approximationSetUnknownGenotypeCombinations}
\title{Approximations for the set of combinations of unknown contributors}
\usage{

  approximationSetUnknownGenotypeCombinations(optimalUnkownProfileCombinationList,
  method = "EA", sampleTibble, H = NULL, potentialParentsList,
  stutterRatioModel = NULL, HArguments = NULL,
  control = optimalUnknownProfileCombination.control())
}
\arguments{
\item{optimalUnkownProfileCombinationList}{An object created by the 'optimalUnknownProfileCombination' function.}

\item{method}{The approximation method, either: 'EA', 'oneStep', or 'MH'.}

\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{H}{A hypothesis.}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample.}

\item{stutterRatioModel}{An \link{lm}-object modelling the stutter ratio given the BLMM. Note: this is only used if 'potentialParentsList' is NULL, but even then it can be NULL itself, implying no stuttering in the sample.}

\item{HArguments}{If 'H' is 'NULL' a list of arguments passed to the 'setHypothesis' function. It must contain 'numberOfContributors', 'knownProfilesList', and 'theta'.}

\item{control}{An \link{optimalUnknownProfileCombination.control} object.}
}
\value{
A list containing the unnormalised probabilitis for every marker, the corresponding genotypes, and approximations to the posterior probabilities for every unknown genotype combination for every marker.
}
\description{
Three approximation to the sum over the set of unknown genotype combinations using: (1) a simple one-step approximation, (2) the set of the fittest individuals, and (3) a Metropolis-Hastings sampler.
}
