% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{collectSamplePopulation}
\alias{collectSamplePopulation}
\title{Collect coverage and population information}
\usage{
collectSamplePopulation(coverageTibble, populationLadder)
}
\arguments{
\item{coverageTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{populationLadder}{A \link{tibble} containing the markers, regions, and allele frequencies of the population.}
}
\value{
A joined tibble, having added any regions not seen in the 'coverageTibble' with a coverage set to 0, and a column of allele frequencies from the 'populationLadder'.
}
\description{
Collects a coverage tibble and the population ladder, by the 'Marker' column.
}
