% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{devianceResidualPG1}
\alias{devianceResidualPG1}
\title{Deviance residuals of the Poisson-gamma distribution of order 1}
\usage{
devianceResidualPG1(x, mean, dispersion)
}
\arguments{
\item{x}{the count.}

\item{mean}{the expected count.}

\item{dispersion}{the overdispersion.}
}
\value{
The deviance residual.
}
\description{
Deviance residuals of the Poisson-gamma distribution of order 1
}
