% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{devianceResidualPoissonGammaDistribution}
\alias{devianceResidualPoissonGammaDistribution}
\title{Deviance residuals of the Poisson-gamma distribution}
\usage{
devianceResidualPoissonGammaDistribution(x, mean, dispersion)
}
\arguments{
\item{x}{the count.}

\item{mean}{the expected count.}

\item{dispersion}{the overdispersion.}
}
\value{
The deviance residual.
}
\description{
Deviance residuals of the Poisson-gamma distribution
}
