% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\docType{data}
\name{examplePopulation}
\alias{examplePopulation}
\title{An example population.}
\usage{
data(examplePopulation)
}
\description{
An example populations generated by the \link{generateAllelicLadder}-function.
}
\keyword{datasets}
