% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{genotypeMatrix}
\alias{genotypeMatrix}
\title{Create genotype matrix}
\usage{
genotypeMatrix(coverageTibble, trueProfiles)
}
\arguments{
\item{coverageTibble}{A tibble containing sample information.}

\item{trueProfiles}{A list of tibbles, an element for each contributor.}
}
\value{
A matrix with the allele counts of of each contributor.
}
\description{
Creates genotype matrix, given a sample and the true profiles of the contributors.
}
