% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotComparingProfiles}
\alias{ggplotComparingProfiles}
\title{Compare profile combinations}
\usage{
ggplotComparingProfiles(sampleTibble, profileList1, profileList2,
  plotDifferencesOnly = FALSE, profileNames = NULL)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{profileList1}{A list containing the expected contribution matrix and the mixture paramters of profile 1.}

\item{profileList2}{A list containing the expected contribution matrix and the mixture paramters of profile 2.}

\item{plotDifferencesOnly}{Restrict the returned \link{ggplot}-object to markers where the two profiles mismatch.}
}
\value{
ggplot2-object.
}
\description{
Compare profile combinations
}
\details{
Function used for visually comparing two profiles.
}
