% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotPPPlotProfilesEnvelopes}
\alias{ggplotPPPlotProfilesEnvelopes}
\title{P-P plot and envelope of a DNA profile}
\usage{
ggplotPPPlotProfilesEnvelopes(sampleTibble, profileList, markerImbalances,
  knownProfilesList, potentialParentsList, stutterRatioModel = NULL,
  levelsOfStutterRecursion = 2, convexMarkerImbalanceInterpolation = 0.8,
  tolerance = rep(1e-08, 4), numberOfThreads = 4,
  numberOfSimulations = 1000, envelopes = c(0.005, 0.995), trace = FALSE,
  traceLimit = 100)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{profileList}{A list containing the expected contribution matrix and paramters of the fitted profile.}

\item{markerImbalances}{A vector of prior marker imbalances.}

\item{knownProfilesList}{A list of tibbles containing the alleles of the known contributors.}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample. If NULL then a stutterRatioModel should be provided.}

\item{stutterRatioModel}{A linear model of class \link{lm} modelling the relationship between coverage and stutter. Only needed if the potential parents list is not provided.}

\item{levelsOfStutterRecursion}{The number of layers used in the stutter recursion.}

\item{convexMarkerImbalanceInterpolation}{A fraction used to create a convex combination of the of MoM and the prior estimates of the marker imbalances.}

\item{tolerance}{The tolerance used for termination in parameter estimation.}

\item{numberOfThreads}{The number of threads passed to the \link{potentialParentsMultiCore}-function if applicable.}

\item{numberOfSimulations}{The number of simulations used to create the envelopes.}

\item{envelopes}{The envelope quantiles.}

\item{trace}{Show trace (TRUE/FALSE)?}

\item{traceLimit}{Limits the trace of the parallel implementation.}
}
\value{
ggplot2-object.
}
\description{
P-P plot and envelopes for the allele coverage component of a DNA profile.
}
