% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotPredictionIntervals}
\alias{ggplotPredictionIntervals}
\title{Prediction intervals.}
\usage{
ggplotPredictionIntervals(sampleTibble, profileList,
  predictionInterval = c(0.005, 0.995), numberOfSimulations = 10000,
  contributorNames = NULL)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{profileList}{A list containing the expected contribution matrix and paramters of the fitted profile.}

\item{predictionInterval}{The upper and lower quantiles of the prediction interval.}

\item{numberOfSimulations}{The number of simulations used to create the prediction interval.}

\item{contributorNames}{The names of the contributors; must be 'NULL', or a vector of length '1' or the number of contributors.}
}
\value{
ggplot2-object.
}
\description{
Prediction intervals of the Poisson-gamma model.
}
