% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotQQPlotProfiles}
\alias{ggplotQQPlotProfiles}
\title{Q-Q plots for profile list}
\usage{
ggplotQQPlotProfiles(sampleTibble, profileList)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{profileList}{A list containing the expected contribution matrix and paramters of the fitted profile.}
}
\value{
ggplot2-object.
}
\description{
Q-Q plots for the allele coverage and noise Poisson-gamma coverage models.
}
