% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{head.setOfUnknownGenotypes}
\alias{head.setOfUnknownGenotypes}
\title{Head of 'setOfUnknownGenotypes'-object}
\usage{
\method{head}{setOfUnknownGenotypes}(x, ...)
}
\arguments{
\item{x}{A list of unknown genotypes, normalised, and unnormalised posterior probabilities.}

\item{...}{Additional arguments: The number of elements, 'n', and the elements of the list to be returned, 'outputElements'.}
}
\value{
A reduced 'setOfUnknownGenotypesList'-object.
}
\description{
Returns the most prevalent elements of 'setOfUnknownGenotypes'-object for each entry in the list.
}
