% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR.control}
\alias{LR.control}
\title{LR control function}
\usage{
LR.control(numberOfPopulations = 4, populationSize = 10,
  numberOfIterations = 25, numberOfInnerIterations = 10,
  numberOfIterationsEqualMinMax = 10, fractionOfPopulationsMax = NULL,
  numberOfFittestIndividuals = 10, parentSelectionWindowSize = 5,
  allowParentSurvival = TRUE, crossoverProbability = NULL,
  mutationProbabilityLowerLimit = NULL, mutationDegreesOfFreedom = 100,
  mutationDecayRate = 2, mutationDecay = NULL,
  fractionFittestIndividuals = 1, hillClimbingDirections = 1,
  hillClimbingIterations = 1, tolerance = 1e-04, seed = NULL,
  trace = TRUE, simplifiedReturn = FALSE, numberOfThreads = 4)
}
\arguments{
\item{numberOfPopulations}{The number of sub-populations.}

\item{populationSize}{The size of the sub-populations.}

\item{numberOfIterations}{The maximum number of (outer) iterations.}

\item{numberOfInnerIterations}{The number of inner iterations.}

\item{fractionOfPopulationsMax}{Fraction of the number of sub-populations with the same maximum needed before convergence counter initiates.}

\item{numberOfFittestIndividuals}{The number of unique individuals stored and returned.}

\item{parentSelectionWindowSize}{The size of the parent selecetion window.}

\item{allowParentSurvival}{Should parents be allowed to survive from iteration to iteration (TRUE/FALSE).}

\item{crossoverProbability}{The cross-over probability.}

\item{mutationProbabilityLowerLimit}{The lower limit to the mutation probability.}

\item{mutationDegreesOfFreedom}{The degrees of freedom of the t-distribution used to create the mutation probabilities.}

\item{mutationDecay}{The decay of the mutation probability to the lower limit.}

\item{hillClimbingDirections}{The number of hill climbing directions (not in use at the moment).}

\item{hillClimbingIterations}{The number of iterations each child or parent is hill climbed.}

\item{tolerance}{Tolerance of internal log-likelihood maximisation.}

\item{seed}{A seed for the c++ implementaion.}

\item{trace}{Show trace (TRUE/FALSE)?}

\item{simplifiedReturn}{Should the returned list be simplified (TRUE/FALSE)?}

\item{numberOfThreads}{The maximum number of threads allowed.}

\item{numberOfIterationsEqualMax}{The number of iterations with the number of sub-populations with the same 'largest' individual divided 'numberOfPopulations' less than or equal to 'fractionOfPopulationsMax'.}
}
\value{
A list containing all relevant parameters.
}
\description{
A function setting all relavent parameters used internally in the \link{LR}-function.
}
\details{
If 'numberOfFittestIndividuals' is 'NULL' (default), then the number of fittest individuals stored across iterations is ceiling(0.1 * 'populationSize'). If the number of populations is larger than 1, then 'populationSize' refers to the sub-population size. Thus, it is recommended this parameter is set if running multiple populations.

If 'fractionOfPopulationsMax' is 'NULL' (default), it is set as the maximum of 0.05 and 1 / 'numberOfPopulations'.

If 'mutationDecay' is 'NULL' (default) and if the number of populations is 1, then it is set as seq(0, by = mutationDecayRate * 4 / (numberOfIterations), length.out = numberOfIterations). If the population size is larger than 1 the parameter 'numberOfIterations' is replaced by 'numberOfIterations * numberOfInnerIterations' i.e. the total number of iterations.

If 'seed' is 'NULL', then it is sampled from uniformly from [1; 10^6].
}
