% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Likelihood ratio}
\usage{
LR(sampleTibble, Hp, Hd, markerImbalances = NULL,
  potentialParentsList = NULL, stutterRatioModel = NULL,
  control = optimalUnknownProfileCombination.control())
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{Hp}{The prosecutor hypothesis (see \link{setHypothesis}).}

\item{Hd}{The defence hypothesis.}

\item{markerImbalances}{A vector of prior marker imbalances.}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample. If NULL then a stutterRatioModel should be provided.}

\item{stutterRatioModel}{A linear model of class \link{lm} modelling the relationship between coverage and stutter. Only needed if the potential parents list is not provided.}

\item{control}{An \link{optimalUnknownProfileCombination.control} object.}
}
\value{
A list of likelihood ratios comparing the two hypotheses (always calculated as Hp / Hd).
}
\description{
Calculates an approximate likelihood ratio given two competing hypotheses (in the future lists of competing hypotheses).
}
