% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{optimalUnknownProfileCombination.control}
\alias{optimalUnknownProfileCombination.control}
\title{Control function for the 'optimalUnknownProfileCombination' function}
\usage{
optimalUnknownProfileCombination.control(..., EAApproximationType = 1,
  numberOfSimulationsMH = 10000, suggestionMH = "random")
}
\arguments{
\item{...}{arguments passed to the \link{LR.control} function.}

\item{EAApproximationType}{The type of EA approximation used to find the normalising constant.}

\item{numberOfSimulationsMH}{The number of simulations used for each marker of the Metropolis-Hastings algorithm.}

\item{suggestionMH}{The proposal distribution used in the Metropolis-Hastings algorithm.}
}
\value{
A list of default parameters.
}
\description{
Control function for the 'optimalUnknownProfileCombination' function
}
\details{
A control function setting default parameters for the \link{optimalUnknownProfileCombination} function. It takes the same parameters as the \link{LR.control} function, with the expection of 'simplifiedReturn'.
}
