% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{optimalUnknownProfileCombination}
\alias{optimalUnknownProfileCombination}
\title{Optimal unknown profile combinations.}
\usage{
optimalUnknownProfileCombination(sampleTibble, markerImbalances, H,
  potentialParentsList, stutterRatioModel = NULL, HArguments = NULL,
  control = optimalUnknownProfileCombination.control())
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{markerImbalances}{The prior marker imbalances.}

\item{H}{A hypothesis.}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample.}

\item{stutterRatioModel}{An \link{lm}-object modelling the stutter ratio given the BLMM. Note: this is only used if 'potentialParentsList' is NULL, but even then it can be NULL itself, implying no stuttering in the sample.}

\item{HArguments}{If 'H' is 'NULL' a list of arguments passed to the 'setHypothesis' function. It must contain 'numberOfContributors', 'knownProfilesList', and 'theta'.}

\item{control}{An \link{optimalUnknownProfileCombination.control} object.}
}
\value{
A list of the unknown profile combinations contributing the most to the probability of the evidence under the provided hypothesis. The size of the list is controlled by 'control$numberOfFittestIndividuals'.
}
\description{
Finds the set of unknown profile combinations, as well as parameters, maximising the log-likehood function.
}
