% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{optimalUnknownProfilesHi}
\alias{optimalUnknownProfilesHi}
\title{Optimal unknown profile combinations under H_i}
\usage{
optimalUnknownProfilesHi(sampleTibble, H, potentialParentsList,
  allKnownProfiles, control)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{H}{A hypothesis (see the \link{setHypothesis}-function for the general structure).}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample.}

\item{control}{An \link{LR.control} object.}
}
\value{
A list of the unknown profile combinations contributing the most to the probability of the evidence under the provided hypothesis. The size of the list is controlled by control$numberOfFittestIndividuals.
}
\description{
Finds the set of unknown profile combinations contributing the most to the probability of the evidence under H_i. Note: this is mostly for use internally in the \link{LR}-function.
}
