% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sampleCoverage}
\alias{sampleCoverage}
\title{Sample MPS coverage}
\usage{
sampleCoverage(trueProfiles, markerImbalances, populationLadder,
  stutterRatioModel, alleleCoverageParameters = list(),
  noiseParameters = list(), p = NULL)
}
\arguments{
\item{trueProfiles}{A list of \link{tibble}'s containing the true profiles of the mixture.}

\item{markerImbalances}{A vector of the marker imbalances.}

\item{populationLadder}{An allelic ladder containing the possible STR regions of each marker and their allele frequencies.}

\item{alleleCoverageParameters}{A list containing the parameters of the allele coverage model.}

\item{noiseParameters}{A list containing the paramters of the noise model.}

\item{p}{A probability threshold for the created stutters.}

\item{numberOfAlleles}{A vector of the total number of alleles per marker.}
}
\value{
A \link{tibble} with the sampled coverage, the marker, and the marker imbalances.
}
\description{
Sample coverage under the Poisson-gamma model of Vilsen et. al 2017
}
