% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{sampleGenotypesHWE}
\alias{sampleGenotypesHWE}
\title{Sample genotypes}
\usage{
sampleGenotypesHWE(numberOfContributors, populationLadder,
  contributorNames = NULL)
}
\arguments{
\item{numberOfContributors}{The number of profiles which should be sampled.}

\item{populationLadder}{A \link{tibble} containing the markers, regions, and allele frequencies of the population.}

\item{contributorNames}{A vector of names. The length should be the number of profiles.}
}
\value{
A list of tibbles, one for each sampled profile.
}
\description{
Sample genotypes, given the number of profiles and a population, assuming the population was in HWE.
}
