% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{estimateParametersOfKnownProfiles}
\alias{estimateParametersOfKnownProfiles}
\title{Estimate parameters}
\usage{
estimateParametersOfKnownProfiles(sampleTibble, knownProfilesList,
  potentialParentsList, stutterRatioModel = NULL, tolerance)
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{knownProfilesList}{A list of tibbles containing the alleles of the known contributors.}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample.}

\item{stutterRatioModel}{An \link{lm}-object modelling the stutter ratio given the BLMM. Note: this is only used if 'potentialParentsList' is NULL, but even then it can be NULL itself, implying no stuttering in the sample.}

\item{tolerance}{The tolerance used for termination in parameter estimation.}
}
\value{
A list of estimated parameters.
}
\description{
Estimate parameters
}
\details{
Estimates the allele and noise parameters in cases where all contirbutors are assumed known.
}
