% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{optimalUnknownProfileCombination.control}
\alias{optimalUnknownProfileCombination.control}
\title{Control function for the 'optimalUnknownProfileCombination' function}
\usage{
optimalUnknownProfileCombination.control(..., numberOfSimulationsMH = 10000,
  suggestionMH = "guided")
}
\value{
A list of default parameters.
}
\description{
Control function for the 'optimalUnknownProfileCombination' function
}
\details{
A control function setting default parameters for the \link{optimalUnknownProfileCombination} function. It takes the same parameters as the \link{LR.control} function, with the expection of 'simplifiedReturn'.
}
