% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{optimalUnknownProfileCombination.control}
\alias{optimalUnknownProfileCombination.control}
\title{Control function for the 'optimalUnknownProfileCombination' function}
\usage{
optimalUnknownProfileCombination.control(numberOfPopulations = 4,
  populationSize = 10, numberOfIterations = 25,
  numberOfInnerIterations = 10, numberOfIterationsEqualMinMax = 10,
  fractionOfPopulationsMax = NULL, numberOfFittestIndividuals = 10,
  parentSelectionWindowSize = 5, allowParentSurvival = TRUE,
  crossoverProbability = NULL, mutationProbabilityLowerLimit = NULL,
  mutationDegreesOfFreedom = 100, mutationDecayRate = 2,
  mutationDecay = NULL, fractionFittestIndividuals = 1,
  hillClimbingDirections = 1, hillClimbingIterations = 1,
  convexMarkerImbalanceInterpolation = 0.8, tolerance = 1e-06,
  seed = NULL, trace = TRUE, numberOfThreads = 4,
  levelsOfStutterRecursion = 2, numberOfSimulationsMH = 10000,
  suggestionMH = "guided")
}
\value{
A list of default parameters.
}
\description{
Control function for the 'optimalUnknownProfileCombination' function
}
\details{
A control function setting default parameters for the \link{optimalUnknownProfileCombination} function. It takes the same parameters as the \link{LR.control} function, with the expection of 'simplifiedReturn'.
}
