% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHypothesisFunctions.R
\name{optimalUnknownProfileCombination}
\alias{optimalUnknownProfileCombination}
\title{Optimal unknown profile combinations.}
\usage{
optimalUnknownProfileCombination(sampleTibble, markerImbalances,
  numberOfContributors, knownProfilesList, theta, potentialParentsList,
  stutterRatioModel = NULL,
  control = optimalUnknownProfileCombination.control())
}
\arguments{
\item{sampleTibble}{A \link{tibble} containing the coverage, the marker, and marker imbalance scalar of each allele in the sample.}

\item{numberOfContributors}{The total number of contributors to the mixture.}

\item{knownProfilesList}{A list of tibbles containing the alleles of the known contributors.}

\item{theta}{The inbreeding coefficient (Fst).}

\item{potentialParentsList}{A list containing a list of potential parents for each allele in the sample.}

\item{stutterRatioModel}{An \link{lm}-object modelling the stutter ratio given the BLMM. Note: this is only used if 'potentialParentsList' is NULL, but even then it can be NULL itself, implying no stuttering in the sample.}

\item{control}{An \link{LR.control} object.}
}
\value{
A list of the unknown profile combinations contributing the most to the probability of the evidence under the provided hypothesis. The size of the list is controlled by 'control$numberOfFittestIndividuals'.
}
\description{
Finds the set of unknown profile combinations, as well as parameters, maximising the log-likehood function.
}
